Move error logic into base
